api,common,transport: fix clippy warnings in async-io code
Fix all clippy warnings reported when building the async-io feature
combinations, so that these combinations can be added to `make check`
with `-Dwarnings`:

- vfs/async_io: remove useless `map(Into::into)` conversion on the
  async open handle
- server/async_io: pass `in_header` directly, it is already a
  reference
- fusedev/mod: use `inspect()` instead of `map()` for side effects
- transport/mod: remove redundant `as usize` casts
- async_runtime: use `is_err()` and implicit final return
- mpmc: add elided lifetime in return type

No functional change.

Related-to: #188

+++ src/transport/mod.rs
@@ -196,7 +196,7 @@ impl<S: BitmapSlice> IoBuffers<'_, S> {
             ));
 
             // Don't need check_sub() as we just made sure rem >= local_buf.len()
-            rem -= local_buf.len() as usize;
+            rem -= local_buf.len();
         }
 
         bufs
@@ -227,7 +227,7 @@ impl<S: BitmapSlice> IoBuffers<'_, S> {
             ));
 
             // Don't need check_sub() as we just made sure rem >= local_buf.len()
-            rem -= local_buf.len() as usize;
+            rem -= local_buf.len();
         }
 
         bufs
